+++ src/app/pushup-counter/page.tsx
@@ -1,28 +1,11 @@
-'use client';
-
-import { useSearchParams } from 'next/navigation';
-import Link from 'next/link';
-import { PushupType } from '@/types/pushup';
-import { PushupDetector } from '@/components/pushup-counter';
-import { PushupTypeSelector } from '@/components/pushup-type-selector';
+import { Suspense } from 'react';
+import PushupCounterView from '@/components/pushup-counter-view';
 
 export default function PushupCounterPage() {
-  const searchParams = useSearchParams();
-  const type = searchParams.get('type');
-
   return (
-    <div className="relative">
-      {/* 헤더 */}
-      <div className="bg-gray-900 text-white flex items-center justify-between px-4 py-3 border-b border-gray-200">
-        <div className="w-6"></div>
-        <h1 className="text-xl font-bold flex-1 text-center">근력 측정</h1>
-        <div className="w-6"></div>
-      </div>
-      {type ? (
-        <PushupDetector type={type as PushupType} />
-      ) : (
-        <PushupTypeSelector />
-      )}
-    </div>
+    <Suspense fallback={<div className="flex items-center justify-center h-screen text-white bg-gray-900">로딩 중...</div>}>
+      <PushupCounterView />
+    </Suspense>
   );
 }
+
